transcription-system/python-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==11.0.3
python-multipart==0.0.6
faster-whisper==1.0.1
numpy==1.24.3
scipy==1.11.4
librosa==0.10.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
torch==2.1.1
torchaudio==2.1.1
